PyBay2018/requirements.txt
grpcio==1.14.1
grpcio-tools==1.14.1
googleapis-common-protos==1.5.3
requests==2.19.1
twisted[h2]==18.7.0
hyper==0.7.0
PyJWT==1.6.4
Django==1.11
